Add a marker file for MyPy
In order for MyPy to use type hints from a package, you have to tell it with a marker file. This will make MyPy more strict in the tests. Otherwise statements like the following are ignored because it considers our types "Any":

var: IntegerValue = "3"  # No error without the py.typed marker file