Add file size as metadata
We never used to have this because the file contents were always immediately brought along with file value instances, so you could just go look at the file on disk. Now that transfers are delayed until you call a method, it would be useful to some clients to know the size before downloading.