Initial commit of ITypeLibrary API
Initial commit to save progress, though this work is not complete yet.
Changed VariableType to allow getting the Python type for a VariableType.
This required throwing an exception, which created a circular dependency on exceptions.py. Changed IncompatibleTypeException to a standard exception and created a factory method to instantiate it from VariableType objects instead.
Craeted ITypeLibrary and the UnifiedTypeLibrary implementation.