Create AbstractInitializerTypeLibrary
This is not a complete commit. Merely saving a point in time for backup.

Creates an AbstractInitializerTypeLibrary that implements the best we can do in Python. Utilizes is_linking_allowed and some other rules (to be documented in a future commit) to validate the conversion, then merely applies the destination type initializer.

Also begins, but does not complete, the work to fix the exception/variable_type circular dependency by bringing back the VaraibleFactory from the DotNet API.

This commit moves IncompatibleTypesException to the API package and renames it per PEP8.